app/Http/Requests/Instructor/Attendanceディレクトリが肥大化しているための整理(yuki)
